requirements.txt
deap
numpy
matplotlib
pillow
requests
